requirements.txt
flask==2.3.3
flask-wtf==1.1.1
requests==2.31.0
tldextract==3.4.4
python-dotenv==1.0.0
